requirements.txt
Django==1.6.10
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.2
argparse==1.2.1
braintree==3.10.0
dj-database-url==0.3.0
django-classy-tags==0.5.2
django-cms==3.0.9
django-common==0.1.51
django-mptt==0.6.1
django-paypal==0.2
django-polymorphic==0.6.1
django-reversion==1.8.5
django-sekizai==0.8.1
django-shop==0.2.0
django-shop-braintree==0.1
django-shop-discounts==0.0.7
django-shop-paypal==0.0.1
django-shop-simplenotifications==0.2
-e git+https://github.com/dokterbob/django-wishlist.git@d3092173f02b840e9271bfeb0538c7f338914535#egg=django_wishlist-master
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.1
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
fudge==1.0.3
html5lib==0.999
ipython==2.4.0
jsonfield==1.0.0
pytz==2014.10
requests==2.5.1
six==1.9.0
wsgiref==0.1.2
